[Prankul] feat . InsuranceApp added Payment functionality

+++ ModelLayer/Models/CommissionModel.cs
@@ -5,6 +5,6 @@ namespace ModelLayer.Models
     public class CommissionModel
     {
         [Required]
-        public long PolicyId { get; set; }
+        public int PolicyId { get; set; }
     }
 }
